Description obligatoire.

+++ workspace/edition/src/main/java/org/osivia/services/workspace/edition/portlet/repository/WorkspaceEditionCommand.java
@@ -140,19 +140,21 @@ public class WorkspaceEditionCommand implements INuxeoCommand {
         }
 
         // Banner
-        Image banner = this.form.getBanner();
-        if (banner.isUpdated()) {
-            // Temporary file
-            File temporaryFile = banner.getTemporaryFile();
-            // File blob
-            Blob blob = new FileBlob(temporaryFile);
+        if (this.form.isRoot()) {
+            Image banner = this.form.getBanner();
+            if (banner.isUpdated()) {
+                // Temporary file
+                File temporaryFile = banner.getTemporaryFile();
+                // File blob
+                Blob blob = new FileBlob(temporaryFile);
 
-            documentService.setBlob(workspace, blob, "ttcs:headImage");
+                documentService.setBlob(workspace, blob, "ttcs:headImage");
 
-            // Delete temporary file
-            temporaryFile.delete();
-        } else if (banner.isDeleted()) {
-            documentService.removeBlob(workspace, "ttcs:headImage");
+                // Delete temporary file
+                temporaryFile.delete();
+            } else if (banner.isDeleted()) {
+                documentService.removeBlob(workspace, "ttcs:headImage");
+            }
         }
     }
